Move PollForm from features/composer to features/compose
