+++ app/javascript/flavours/glitch/features/compose/components/poll_form.js
