+++ app/javascript/flavours/glitch/features/compose/containers/poll_form_container.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import PollForm from './components/poll_form';
-import { addPollOption, removePollOption, changePollOption, changePollSettings } from '../../../actions/compose';
+import PollForm from '../components/poll_form';
+import { addPollOption, removePollOption, changePollOption, changePollSettings } from 'flavours/glitch/actions/compose';
 
 const mapStateToProps = state => ({
   options: state.getIn(['compose', 'poll', 'options']),
